Add files via upload
This Java program contains the code to resolve knapsack problem. It first takes the total weight required. Then, it takes the total number of weights used in the problem, and finally, it takes all the weight which is used to calculate the sum equal or closer to the total weight.